Fix a crash in the rqt_py_console dialog box. (#15)
This commit fixes an issue when you click on the "gear"
icon in the rqt_py_console.  That pops up a dialog box,
but if you click "Cancel" there it currently crashes.
That's because clicking Cancel returns None, but then
we subsequently try to subscript it.  If we get a None
out of the dialog, just skip any further processing.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>